Move prefab to service level
