+++ Runtime/Definitions/BaseCameraServiceModuleProfile.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) Reality Collective. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
-using RealityCollective.Attributes;
 using RealityCollective.ServiceFramework.Definitions;
 using RealityToolkit.CameraService.Interfaces;
 using System;
@@ -16,14 +15,6 @@ namespace RealityToolkit.CameraService.Definitions
     public class BaseCameraServiceModuleProfile : BaseProfile
     {
         [Header("General")]
-        [SerializeField, Tooltip("The camera rig prefab used by this service module."), Prefab(typeof(ICameraRig))]
-        private GameObject rigPrefab = null;
-
-        /// <summary>
-        /// The camera rig prefab used by this service module.
-        /// </summary>
-        public GameObject RigPrefab => rigPrefab;
-
         [SerializeField]
         [Tooltip("Sets the tracking type of the camera.")]
         private TrackingType trackingType = TrackingType.Auto;
